+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 5.0.0
+* [feature] animate collapse/expand with only one parameter `panelAnimationDuration`
+* [feature] optional parameter `panelAnimationCurve` to control the animation
+* [break change] no longer notify precedingScrollExtent to avoid potential notification storms during scrolling
+
+## 4.0.1
+* [fix] - status wrong when panelController.expandPanel called
+
 ## 4.0.0
 * [enhance] - enhance StickyCollapsablePanelController for command style control of collapse/expand the panel
 * [break change] - move parameter `disableCollapsable` and `defaultExpanded` to panelController